Osasto Kuikka Website v0.0.1 Frontend.Router.Helpers

Module with named helpers generated from Frontend.Router.

Functions

forum_path(conn_or_endpoint, action)

forum_path(conn_or_endpoint, action, params)

forum_path(conn_or_endpoint, action, id, params)

forum_url(conn_or_endpoint, action)

forum_url(conn_or_endpoint, action, params)

forum_url(conn_or_endpoint, action, id, params)

home_path(conn_or_endpoint, action)

home_path(conn_or_endpoint, action, params)

home_url(conn_or_endpoint, action)

home_url(conn_or_endpoint, action, params)

member_path(conn_or_endpoint, action)

member_path(conn_or_endpoint, action, params)

member_path(conn_or_endpoint, action, id, params)

member_url(conn_or_endpoint, action)

member_url(conn_or_endpoint, action, params)

member_url(conn_or_endpoint, action, id, params)

path(data, path)

Generates the path information including any necessary prefix.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

steamex_auth_return_to_path(conn_or_endpoint, action)

steamex_auth_return_to_path(conn_or_endpoint, action, params)

steamex_auth_return_to_url(conn_or_endpoint, action)

steamex_auth_return_to_url(conn_or_endpoint, action, params)

url(data)

Generates the connection/endpoint base URL without any path information.