kuikka_website v0.0.1 KuikkaDB

KuikkaDB includes all functions to needed for website to access to database

Functions

get_all_users()

Get all users from kuikkadb

get_user(steamid)

Get one user from kuikkadb with steam id

new_role(name, description)

Add new role to kuikkadb with name and description

new_user(steamid)

Add new user to kuikkadb with steam id

query(query)

Send custom query to kuikkadb

update_user_role(steamid, rolename)

Update user role to kuikkadb with steam id and role name