kuikka_website v0.0.1 KuikkaDB.Controller

Controller containing database functions for user

Functions

get_all_users()

Get all user from database.

This function returns user structs so every user info have been requested steam api.

get_user(steamid)

Get single user from database as user struct.

This function will automatically request data from steam api to build proper user struct for frontend usage.

new_role(name, description)

Insert new role to kuikkadb. Returns error tuple if role already exists otherwise it will return ok tuple when new role was inserted to kuikkadb

new_user(steamid)

Adds new user to database. Should be automatically used when user first time logins from steam.

update_user_role(steamid, rolename)

Update user role to database. This is will return atom with :ok when user was properly updated and {:error, msg} when there is issue with updating.