Kuikka Website v0.0.1 KuikkaDB.Tasks

Tasks for release build of kuikkadb are included in this module

These tasks are meant to run on distillery hooks and are not for use in other parts for kuikkadb

Functions

migrate()

This task will run migrations for kuikkadb.

Example for distillery hook script

echo "Running migrations for kuikkadb"
bin/my-app rpc Elixir.KuikkaDB.Tasks migrate
echo "Migrations run successfully for kuikkadb"